+++ lib/restrictStringCount.ts
